requirements.txt
flask
gunicorn
psycopg2-binary
validators
Flask-Limiter
yarl
bcrypt